requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
flask-socketio>=5.0.0
PyJWT>=2.0.0
requests>=2.25.0
opencv-python>=4.5.0
numpy>=1.19.0
ultralytics>=8.0.0
python-socketio>=5.0.0
eventlet>=0.30.0
